the ati functions need to be _stdcall or they crash
not sure how it works on gcc though. Also this will give a pile of warning as the empty function stub doesn't take params so can't be std_call which means its a different type and it will complain. Assigning a function with a different number of params is bad anyway, however its in the "it failed" case which is just bad either way.