updated Demo Timer and fixed demo crash
Timer is now *2 before starting the demo ( Title screen is always 60fps )
script loading wasn't setting demo to 0 so when it goes to find the first demo it can get the 1st level which has no demo data and it goes bang. Or you get lucky and it has a zero in it, so its "works", now it always has zero.